+++ src/dropdown-menu/_example/dropdown-menu.json
@@ -1,6 +1,8 @@
 {
   "navigationBarTitleText": "Dropdown Menu 下拉菜单",
   "usingComponents": {
+    "t-dropdown-menu": "tdesign-miniprogram/dropdown-menu/dropdown-menu",
+    "t-dropdown-item": "tdesign-miniprogram/dropdown-item/dropdown-item",
     "single": "./single",
     "multi": "./multi"
   }
