+++ src/dropdown-menu/_example/multi/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-dropdown-menu": "tdesign-miniprogram/dropdown-menu/dropdown-menu",
-    "t-dropdown-item": "tdesign-miniprogram/dropdown-menu/dropdown-item"
+    "t-dropdown-item": "tdesign-miniprogram/dropdown-item/dropdown-item"
   }
 }
