+++ src/dropdown-menu/index.ts
@@ -1,6 +1,4 @@
 export * from './type';
 export * from './props';
-export * from './dropdown-item-props';
 
 export * from './dropdown-menu';
-export * from './dropdown-item';
