--- src/dropdown-menu/util.wxs
@@ -1,16 +1,0 @@
-var getTreeClass = function (level) {
-  if (level === 0) return 'leaf';
-  if (level === 1) return 'parent';
-  return 'ancestor';
-};
-
-var getDropdownItemStyles = function (top, zIndex, customStyle) {
-  var topStyle = top ? 'top:' + top + 'px;' : '';
-  var zIndexStyle = zIndex ? 'z-index:' + zIndex + ';' : '';
-  return topStyle + zIndexStyle + customStyle;
-};
-
-module.exports = {
-  getTreeClass: getTreeClass,
-  getDropdownItemStyles: getDropdownItemStyles,
-};
